word spotting initialized

+++ spnt-chart/src/main/java/org/spantus/chart/ChartDescriptionResolver.java
@@ -1,5 +1,5 @@
 package org.spantus.chart;
 
 public interface ChartDescriptionResolver {
-	ChartDescriptionInfo resolve(float time, float value);
+	ChartDescriptionInfo resolve(Long time, float value);
 }
